src: app: config: Add configuration for `git send-email` options
Actions that involve sending email using `git send-email` (currently,
just replying a patchset with the `Reviewed-by` tag) use the hardcoded
options `--dry-run` and `--suppress-cc=all`.

To `patch-hub` more configurable, add configuration for the options of
`git send-email`. The configuration is a simple string that should
contain all the options and values. The configuration name is
`git_send_email_options` and is stored in the JSON that centralizes
`patch-hub` configurations. It is overridable through the
`PATCH_HUB_GIT_SEND_EMAIL_OPTIONS` environment variable (the default
remains `--dry-run` and `--suppress-cc=all`).

Currently, the only way to set this config is through directly editing
the config file or through the environment variable, but the plan is to
add a screen to do this through the application itself.

Note: There are no checks for the validity of the options, but this
doesn't crash the application, only causes the `git send-email` command
to fail.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>