refactor(lore): enforce `mod.rs` style to `lore`
In a recent commit, we introduced the "older" file path style of
`mod.rs`, i.e., for a module, say, `foo::bar`, the file that defines the
module would be located at `src/foo/bar/mod.rs`. However, the submodules
of `lore` still used the "newer" style.

Even though Rust allows us to mix both styles in the same project, we
should stick to one and standardize it, so enforce the `mod.rs` style in
in all submodules of `lore`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>